requirements.txt
flask 
grpcio 
grpcio-tools
redis
geohash2
flask-cors